make coverity happy!

+++ ai/default/daidomestic.cpp
@@ -387,7 +387,7 @@ static void dai_choose_trade_route(struct ai_type *ait, struct city *pcity,
    * For optimal performance AI should build more caravans, but
    * we want it to build less valued buildings too. */
   trader_trait = ai_trait_get_value(TRAIT_TRADER, pplayer);
-  want /= (130 * TRAIT_DEFAULT_VALUE / trader_trait);
+  want /= (130.f * TRAIT_DEFAULT_VALUE / trader_trait);
 
   /* Increase want for trade routes if our economy is very weak.
    * We may have enough gold, but only because we have already set
@@ -587,7 +587,7 @@ struct adv_choice *domestic_advisor_choose_build(struct ai_type *ait,
     building_advisor_choose(pcity, cur);
     cur->want = cur->want
                 * (0.5
-                   + (ai_trait_get_value(TRAIT_BUILDER, pplayer)
+                   + double((ai_trait_get_value(TRAIT_BUILDER, pplayer))
                       / TRAIT_DEFAULT_VALUE / 2));
     choice = adv_better_choice_free(choice, cur);
