+++ ai/default/daimilitary.cpp
@@ -473,7 +473,7 @@ static void dai_reevaluate_building(struct city *pcity, adv_want *value,
   if (urgency > 0 && danger > defense * 2) {
     *value += 100;
   } else if (defense != 0 && danger > defense) {
-    *value = MAX(danger * 100 / defense, *value);
+    *value = MAX(danger * 100.f / defense, *value);
   }
 }
 
@@ -1460,10 +1460,11 @@ static void adjust_ai_unit_choice(struct city *pcity,
           != B_LAST
       && !city_has_building(pcity, improvement_by_number(id))) {
     choice->value.building = improvement_by_number(id);
-    choice->want = choice->want
-                   * (0.5
-                      + (ai_trait_get_value(TRAIT_BUILDER, city_owner(pcity))
-                         / TRAIT_DEFAULT_VALUE / 2));
+    choice->want =
+        choice->want
+        * (0.5
+           + double((ai_trait_get_value(TRAIT_BUILDER, city_owner(pcity)))
+                    / TRAIT_DEFAULT_VALUE / 2));
     choice->type = CT_BUILDING;
     adv_choice_set_use(choice, "veterancy building");
   }
@@ -1591,10 +1592,11 @@ struct adv_choice *military_advisor_choose_build(
           choice->value.building = pimprove;
           /* building_want is hacked by assess_danger */
           choice->want = pcity->server.adv->building_want[wall_id];
-          choice->want = choice->want
-                         * (0.5
-                            + (ai_trait_get_value(TRAIT_BUILDER, pplayer)
-                               / TRAIT_DEFAULT_VALUE / 2));
+          choice->want =
+              choice->want
+              * (0.5
+                 + double((ai_trait_get_value(TRAIT_BUILDER, pplayer))
+                          / TRAIT_DEFAULT_VALUE / 2));
           if (urgency == 0 && choice->want > 100) {
             choice->want = 100;
           }
